fix: добавил методы FilmLikes в FilmStorage, лишнее убрал.
