+++ src/main/java/ru/yandex/practicum/filmorate/storage/film/FilmStorage.java
@@ -12,4 +12,10 @@ public interface FilmStorage {
     Film updateFilm(Film film);
 
     Film getFilmById(long filmId);
+
+    void addLike(long filmId, long userId);
+
+    void deleteLike(long filmId, long userId);
+
+    Collection<Film> getPopularFilms(int count);
 }
